[update] api 분리 리팩토링
